+++ jest.config.js
@@ -1,11 +1,9 @@
 module.exports = {
-	"roots": [
-		"<rootDir>/src"
-	],
-	transform: {
-		"^.+\\.tsx?$": "ts-jest"
-	},
-	verbose: true,
-	collectCoverage: true,
-	coverageReporters: ["text-summary"]
-}
+  roots: ["<rootDir>/src"],
+  transform: {
+    "^.+\\.tsx?$": "ts-jest",
+  },
+  verbose: true,
+  collectCoverage: true,
+  coverageReporters: ["text-summary", "html"],
+};