+++ src/MIDIval.ts
@@ -1,26 +1,22 @@
-import {MIDIValInput} from "./MIDIValInput";
-import {MIDIValOutput} from "./MIDIValOutput";
+import { MIDIValInput } from "./MIDIValInput";
+import { MIDIValOutput } from "./MIDIValOutput";
 
-import {IMIDIInput} from "./wrappers/inputs/IMIDIInput";
+import { IMIDIInput } from "./wrappers/inputs/IMIDIInput";
 
-import {MIDIAccess} from "./wrappers/access/MIDIAccess";
-import {IMIDIOutput} from "./wrappers/outputs/IMIDIOutput";
-import {IMIDIAccess} from "./wrappers/access/IMIDIAccess";
-import { CallbackType, Omnibus, UnregisterCallback } from "@hypersphere/omnibus";
-
-type EventDefinitions = {
-  inputDeviceConnected: [IMIDIInput];
-  inputDeviceDisconnected: [IMIDIInput];
-  outputDeviceConnected: [IMIDIOutput];
-  outputDeviceDisconnected: [IMIDIOutput];
-};
+import { IMIDIOutput } from "./wrappers/outputs/IMIDIOutput";
+import { IMIDIAccess } from "./wrappers/access/IMIDIAccess";
+import { CallbackType, UnregisterCallback } from "@hypersphere/omnibus";
+import { BrowserMIDIAccess } from "./wrappers/access/BrowserMIDIAccess";
 
 export interface ConfigScheme {
-  name?: string | RegExp,
-  manufacturer?: string | RegExp,
+  name?: string | RegExp;
+  manufacturer?: string | RegExp;
 }
 
-export const matchesConfig = (input: IMIDIInput | IMIDIOutput, scheme: ConfigScheme): boolean => {
+export const matchesConfig = (
+  input: IMIDIInput | IMIDIOutput,
+  scheme: ConfigScheme
+): boolean => {
   return Object.keys(scheme).reduce((acc, key) => {
     const val = scheme[key] as string | RegExp;
     if (typeof val === "string") {
@@ -29,12 +25,10 @@ export const matchesConfig = (input: IMIDIInput | IMIDIOutput, scheme: ConfigSch
       return acc && val.test(input[key]);
     }
   }, true);
-}
+};
 
 export class MIDIVal {
-  private static omnibus = new Omnibus<EventDefinitions>();
-
-  private static isSetup: boolean = false;
+  private static isSetupComplete: boolean = false;
   private static accessObject: IMIDIAccess;
 
   /**
@@ -42,52 +36,54 @@ export class MIDIVal {
    * @param newAccess Implementation of IMIDIAccess to be used to provide MIDI objects.
    */
   public static configureAccessObject(newAccess: IMIDIAccess): void {
-    this.isSetup = false;
+    this.isSetupComplete = false;
     this.accessObject = newAccess;
   }
 
   /**
    * Calls callback on every input device that gets connected.
-   * @param fn Callback to be registered
+   * @param callback Callback to be registered
    * @param callOnAlreadyConnected If set to true, the function will instantly trigger for all already connected devices. Default to false
    * @returns Promise resolving to unregister callback when finishes.
    */
-  public static async onInputDeviceConnected(fn: CallbackType<[IMIDIInput]>, callOnAlreadyConnected: boolean = false): Promise<UnregisterCallback> {
-    await this.setupDeviceWatchers();
-    const unregister = this.omnibus.on('inputDeviceConnected', fn);
-
-    if (!callOnAlreadyConnected) {
-      return unregister;
+  public static async onInputDeviceConnected(
+    callback: CallbackType<[IMIDIInput]>,
+    callOnAlreadyConnected: boolean = false
+  ): Promise<UnregisterCallback> {
+    if (callOnAlreadyConnected) {
+      this.accessObject.inputs.forEach(callback);
     }
-
-    // rework later.
-    for (const input of this.accessObject.inputs) {
-      fn(input);
-    }
-
-    return unregister;
+    return this.accessObject.onInputConnected(callback);
   }
-/**
- * Listens to all input devices with a certain config (like name or manufacturer). Configuration can be provided as a string or regex. The callback unlike `onInputDeviceConnected` accepts MIDIValInput. `onInputDeviceConnected` is suitable when you want to filter devices yourself, beyond this configuration object so it does not instantiate objects.
- * @param config Configuration object used to match with device connected
- * @param fn Callback on connection. Connection is already wrapped in MIDIValInput object
- * @returns Promise for Unregister Callback
- */
-  public static async onInputDeviceWithConfigConnected(config: ConfigScheme, fn: (input: MIDIValInput) => void, callOnAlreadyConnected: boolean = false): Promise<UnregisterCallback> {
+  /**
+   * Listens to all input devices with a certain config (like name or manufacturer). Configuration can be provided as a string or regex. The callback unlike `onInputDeviceConnected` accepts MIDIValInput. `onInputDeviceConnected` is suitable when you want to filter devices yourself, beyond this configuration object so it does not instantiate objects.
+   * @param config Configuration object used to match with device connected
+   * @param fn Callback on connection. Connection is already wrapped in MIDIValInput object
+   * @returns Promise for Unregister Callback
+   */
+  public static async onInputDeviceWithConfigConnected(
+    config: ConfigScheme,
+    fn: (input: MIDIValInput) => void,
+    callOnAlreadyConnected: boolean = false
+  ): Promise<UnregisterCallback> {
     return this.onInputDeviceConnected((input) => {
       if (matchesConfig(input, config)) {
         fn(new MIDIValInput(input));
       }
     }, callOnAlreadyConnected);
   }
-  
-/**
- * Listens to all output devices with a certain config (like name or manufacturer). Configuration can be provided as a string or regex. The callback unlike `onOutputDeviceConnected` accepts MIDIValOutput. `onOutputDeviceConnected` is suitable when you want to filter devices yourself, beyond this configuration object so it does not instantiate objects.
- * @param config Configuration object used to match with device connected
- * @param fn Callback on connection. Connection is already wrapped in MIDIValOutput object
- * @returns Promise for Unregister Callback
- */
-  public static async onOutputDeviceWithConfigConnected(config: ConfigScheme, fn: (output: MIDIValOutput) => void, callOnAlreadyConnected: boolean = false): Promise<UnregisterCallback> {
+
+  /**
+   * Listens to all output devices with a certain config (like name or manufacturer). Configuration can be provided as a string or regex. The callback unlike `onOutputDeviceConnected` accepts MIDIValOutput. `onOutputDeviceConnected` is suitable when you want to filter devices yourself, beyond this configuration object so it does not instantiate objects.
+   * @param config Configuration object used to match with device connected
+   * @param fn Callback on connection. Connection is already wrapped in MIDIValOutput object
+   * @returns Promise for Unregister Callback
+   */
+  public static async onOutputDeviceWithConfigConnected(
+    config: ConfigScheme,
+    fn: (output: MIDIValOutput) => void,
+    callOnAlreadyConnected: boolean = false
+  ): Promise<UnregisterCallback> {
     return this.onOutputDeviceConnected((output) => {
       if (matchesConfig(output, config)) {
         fn(new MIDIValOutput(output));
@@ -97,46 +93,40 @@ export class MIDIVal {
 
   /**
    * Calls callback on every output device that gets connected.
-   * @param fn Callback to be registered
+   * @param callback Callback to be registered
    * @param callOnAlreadyConnected If set to true, the function will instantly trigger for all already connected devices. Default to false
    * @returns Promise resolving to unregister callback when finishes.
    */
-   public static async onOutputDeviceConnected(fn: CallbackType<[IMIDIOutput]>, callOnAlreadyConnected: boolean = false): Promise<UnregisterCallback> {
-    await this.setupDeviceWatchers();
-    const unregister = this.omnibus.on('outputDeviceConnected', fn);
-
-    if (!callOnAlreadyConnected) {
-      return unregister;
+  public static async onOutputDeviceConnected(
+    callback: CallbackType<[IMIDIOutput]>,
+    callOnAlreadyConnected: boolean = false
+  ): Promise<UnregisterCallback> {
+    if (callOnAlreadyConnected) {
+      this.accessObject.outputs.forEach(callback);
     }
-
-    // rework later.
-    for (const output of this.accessObject.outputs) {
-      fn(output);
-    }
-
-    return unregister;
+    return this.accessObject.onOutputConnected(callback);
   }
 
   /**
    * Regusters callback on an event of input device being disconnected.
-   * @param fn Callback to be called.
+   * @param callback Callback to be called.
    * @returns promise resolving to unregister callback
    */
-  public static async onInputDeviceDisconnected(fn: CallbackType<[IMIDIInput]>): Promise<UnregisterCallback> {
-    const unregister = this.omnibus.on('inputDeviceConnected', fn);
-    await this.setupDeviceWatchers();
-    return unregister;
+  public static async onInputDeviceDisconnected(
+    callback: CallbackType<[IMIDIInput]>
+  ): Promise<UnregisterCallback> {
+    return this.accessObject.onInputDisconnected(callback);
   }
 
   /**
    * Regusters callback on an event of input device being disconnected.
-   * @param fn Callback to be called.
+   * @param callback Callback to be called.
    * @returns promise resolving to unregister callback
    */
-   public static async onOutputDeviceDisconnected(fn: CallbackType<[IMIDIOutput]>): Promise<UnregisterCallback> {
-    const unregister = this.omnibus.on('outputDeviceDisconnected', fn);
-    await this.setupDeviceWatchers();
-    return unregister;
+  public static async onOutputDeviceDisconnected(
+    callback: CallbackType<[IMIDIOutput]>
+  ): Promise<UnregisterCallback> {
+    return this.accessObject.onOutputDisconnected(callback);
   }
 
   /**
@@ -167,33 +157,15 @@ export class MIDIVal {
   }
 
   private static async setupDeviceWatchers() {
-    if (this.isSetup) {
+    if (this.isSetupComplete) {
       return;
     }
-    this.isSetup = true;
+    this.isSetupComplete = true;
 
     if (!this.accessObject) {
-      this.accessObject = new MIDIAccess();
+      this.accessObject = new BrowserMIDIAccess();
     }
 
-    const access = await this.accessObject.connect();
-
-    for (const input of this.accessObject.inputs) {
-      this.omnibus.trigger('inputDeviceConnected', input);
-    }
-    for (const output of this.accessObject.outputs) {
-      this.omnibus.trigger('outputDeviceConnected', output);
-    }
-
-    
-
-    // FIXME: move somewhere else....
-    // access.onstatechange = (e: WebMidi.MIDIConnectionEvent) => {
-    //   if (e.port.state === "connected") {
-    //     this.staticBuses.deviceConnected.trigger(e.port);
-    //   } else {
-    //     this.staticBuses.deviceDisconnected.trigger(e.port);
-    //   }
-    // };
+    await this.accessObject.connect();
   }
 }