+++ src/errors.ts
@@ -1,13 +1,11 @@
 export class MIDIValError extends Error {
-    constructor(message: string) {
-        super(message);
-    }
+  constructor(message: string) {
+    super(message);
+  }
 
-    get name() {
-        return this.constructor.name;
-    }
+  get name() {
+    return this.constructor.name;
+  }
 }
 
-export class MIDIValConfigurationError extends MIDIValError {
-    
-}
+export class MIDIValConfigurationError extends MIDIValError {}