+++ src/index.ts
@@ -1,10 +1,14 @@
-import {MIDIVal} from "./MIDIval";
-import {MIDIValInput, MIDIValInputOptions} from "./MIDIValInput";
-import {MIDIValOutput} from "./MIDIValOutput";
-import {IMIDIInput} from "./wrappers/inputs/IMIDIInput";
-import {IMIDIOutput} from "./wrappers/outputs/IMIDIOutput";
-import {IMIDIAccess} from "./wrappers/access/IMIDIAccess";
-import { ControlChangeMessage, NoteMessage, ProgramChangeMessage } from "./types/messages";
+import { MIDIVal } from "./MIDIval";
+import { MIDIValInput, MIDIValInputOptions } from "./MIDIValInput";
+import { MIDIValOutput } from "./MIDIValOutput";
+import { IMIDIInput } from "./wrappers/inputs/IMIDIInput";
+import { IMIDIOutput } from "./wrappers/outputs/IMIDIOutput";
+import { IMIDIAccess } from "./wrappers/access/IMIDIAccess";
+import {
+  ControlChangeMessage,
+  NoteMessage,
+  ProgramChangeMessage,
+} from "./types/messages";
 import { MidiMessage } from "./utils/MIDIMessageConvert";
 import { CallbackType, UnregisterCallback } from "@hypersphere/omnibus";