+++ src/utils/MIDIMessageConvert.ts
@@ -7,7 +7,7 @@ export interface MidiMessage {
   command: number;
   data1: number;
   data2: number;
-};
+}
 
 type MidiData = Uint8Array;
 export const isChannelMode = (channel: number): boolean => channel >= 120;
@@ -21,7 +21,11 @@ export const makeMessage = ({
   return Uint8Array.from([command + (channel - 1), data1, data2]);
 };
 
-export const toMidiMessage = ([status, data1, data2]: MidiData): MidiMessage => {
+export const toMidiMessage = ([
+  status,
+  data1,
+  data2,
+]: MidiData): MidiMessage => {
   const command: number = (status >> 4) << 4; // cleaning lower bits
   const channel: number = status - command + 1;
   return {