+++ src/wrappers/access/IMIDIAccess.ts
@@ -1,19 +1,19 @@
 import { CallbackType, UnregisterCallback } from "@hypersphere/omnibus";
-import {IMIDIInput} from "../inputs/IMIDIInput";
-import {IMIDIOutput} from "../outputs/IMIDIOutput";
+import { IMIDIInput } from "../inputs/IMIDIInput";
+import { IMIDIOutput } from "../outputs/IMIDIOutput";
 
 export enum STATUS {
   CONNECTED = "connected",
   DISCONNECTED = "disconnected",
 }
 
-export type InputStateChangeCallback = CallbackType<[IMIDIInput]>; 
+export type InputStateChangeCallback = CallbackType<[IMIDIInput]>;
 export type OutputStateChangeCallback = CallbackType<[IMIDIOutput]>;
 
 export interface IMIDIAccess {
   connect(): Promise<void>;
-  inputs: IMIDIInput[];
-  outputs: IMIDIOutput[];
+  get inputs(): IMIDIInput[];
+  get outputs(): IMIDIOutput[];
   onInputConnected(callback: InputStateChangeCallback): UnregisterCallback;
   onInputDisconnected(callback: InputStateChangeCallback): UnregisterCallback;
   onOutputConnected(callback: OutputStateChangeCallback): UnregisterCallback;