+++ src/wrappers/access/MockMIDIAccess.ts
@@ -3,29 +3,40 @@ import {
   InputStateChangeCallback,
   OutputStateChangeCallback,
 } from "./IMIDIAccess";
-import {MockMIDIOutput} from "../outputs/MockMIDIOutput";
+import { MockMIDIOutput } from "../outputs/MockMIDIOutput";
 import { MidiDeviceProps, MockMIDIInput } from "../inputs/MockMIDIInput";
 import { UnregisterCallback } from "../..";
+import { Omnibus } from "@hypersphere/omnibus";
+
+interface Events {
+  input_connected: [MockMIDIInput];
+  input_disconnected: [MockMIDIInput];
+  output_connected: [MockMIDIOutput];
+  output_disconnected: [MockMIDIOutput];
+}
 
 export class MockMIDIAccess implements IMIDIAccess {
   private mockInputs: MockMIDIInput[];
   private mockOutputs: MockMIDIOutput[];
+  private bus: Omnibus<Events> = new Omnibus<Events>();
 
   constructor() {
     this.mockInputs = [];
     this.mockOutputs = [];
   }
   onInputConnected(callback: InputStateChangeCallback): UnregisterCallback {
-    throw new Error("Method not implemented.");
+    return this.bus.on("input_connected", callback);
   }
   onInputDisconnected(callback: InputStateChangeCallback): UnregisterCallback {
-    throw new Error("Method not implemented.");
+    return this.bus.on("input_disconnected", callback);
   }
   onOutputConnected(callback: OutputStateChangeCallback): UnregisterCallback {
-    throw new Error("Method not implemented.");
+    return this.bus.on("output_connected", callback);
   }
-  onOutputDisconnected(callback: OutputStateChangeCallback): UnregisterCallback {
-    throw new Error("Method not implemented.");
+  onOutputDisconnected(
+    callback: OutputStateChangeCallback
+  ): UnregisterCallback {
+    return this.bus.on("output_disconnected", callback);
   }
 
   async connect(): Promise<void> {
@@ -43,12 +54,24 @@ export class MockMIDIAccess implements IMIDIAccess {
   addInput(props: MidiDeviceProps): MockMIDIInput {
     const input = new MockMIDIInput(props);
     this.mockInputs.push(input);
+    this.bus.trigger("input_connected", input);
     return input;
+  }
+
+  removeInput(device: MockMIDIInput) {
+    this.mockInputs = this.mockInputs.filter((x) => x !== device);
+    this.bus.trigger("input_disconnected", device);
   }
 
   addOutput(props: MidiDeviceProps): MockMIDIOutput {
     const output = new MockMIDIOutput(props);
     this.mockOutputs.push(output);
+    this.bus.trigger("output_connected", output);
     return output;
+  }
+
+  removeOutput(device: MockMIDIOutput) {
+    this.mockOutputs = this.mockOutputs.filter((x) => x !== device);
+    this.bus.trigger("output_disconnected", device);
   }
 }