+++ src/wrappers/outputs/IMIDIOutput.ts
@@ -1,7 +1,6 @@
 export interface IMIDIOutput {
   send(data: Uint8Array | number[]): void;
-  id: string;
-  name: string;
-
-  manufacturer: string;
+  get id(): string;
+  get name(): string;
+  get manufacturer(): string;
 }